ldcg-hpc-tensorflow: add Debian version of job
Both CentOS and Debian jobs use same Ansible playbook and build/publish
scripts.

Debian version requested by Tom Gall in HPCAI-466

+++ ldcg-hpc-tensorflow/build.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
-# first let update system
-sudo dnf -y distrosync
-sudo dnf -y install centos-release-ansible-29
-sudo dnf -y install ansible git which wget
+if [ -e /etc/debian_version ]; then
+    echo "deb http://deb.debian.org/debian/ buster-backports main" | sudo tee /etc/apt/sources.list.d/backports.list
+    sudo apt update
+    sudo apt upgrade -y
+    sudo apt install -y ansible/buster-backports
+else
+    sudo dnf -y distrosync
+    sudo dnf -y install centos-release-ansible-29
+    sudo dnf -y install ansible git
+fi
 
 git clone --depth 1 https://git.linaro.org/ci/job/configs.git
